Consolidate build_image_caption_sd_dataloader(...) and build_image_caption_sdxl_dataloader(...).
