+++ configs/finetune_lora_sdxl_pokemon_1x24gb_example.yaml
@@ -21,7 +21,7 @@ optimizer:
     safeguard_warmup: True
 
 data_loader:
-  type: IMAGE_CAPTION_SDXL_DATA_LOADER
+  type: IMAGE_CAPTION_SD_DATA_LOADER
   dataset:
     type: HF_HUB_IMAGE_CAPTION_DATASET
     dataset_name: lambdalabs/pokemon-blip-captions
