+++ configs/finetune_lora_sdxl_pokemon_1x8gb_example.yaml
@@ -22,7 +22,7 @@ optimizer:
     safeguard_warmup: True
 
 data_loader:
-  type: IMAGE_CAPTION_SDXL_DATA_LOADER
+  type: IMAGE_CAPTION_SD_DATA_LOADER
   dataset:
     type: HF_HUB_IMAGE_CAPTION_DATASET
     dataset_name: lambdalabs/pokemon-blip-captions
