+++ src/invoke_training/config/pipelines/finetune_lora_config.py
@@ -7,7 +7,6 @@ from invoke_training.config.shared.data.data_loader_config import (
     DreamboothSDDataLoaderConfig,
     DreamboothSDXLDataLoaderConfig,
     ImageCaptionSDDataLoaderConfig,
-    ImageCaptionSDXLDataLoaderConfig,
 )
 from invoke_training.config.shared.optimizer.optimizer_config import OptimizerConfig
 
@@ -146,7 +145,7 @@ class FinetuneLoRASDXLConfig(LoRATrainingConfig):
     type: Literal["FINETUNE_LORA_SDXL"] = "FINETUNE_LORA_SDXL"
     optimizer: OptimizerConfig
     data_loader: Annotated[
-        Union[ImageCaptionSDXLDataLoaderConfig, DreamboothSDXLDataLoaderConfig], Field(discriminator="type")
+        Union[ImageCaptionSDDataLoaderConfig, DreamboothSDXLDataLoaderConfig], Field(discriminator="type")
     ]
 
     vae_model: Optional[str] = None
